access-client agentToData is not exported, and is commented
